refactoring of component names to resemble the tab labels
